Converting healing brush to use HAT also offered a headless variant for violajones example
